Fix WaitGroup usage
Calling wg.Done() does not "cancel" the WaitGroup, it does introduce a
bug in which depending on timing the WaitGroup counter might go
negative, what should never happen.

Would be nice to rewrite the method to be correct concurrent code, but
for now should be enough to avoid the panic condition and note that
there is a potential leak.

---

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1370265