feat: Default to Podman Socket
Update s2i's defaults to inspect the host system for Podman sockets.
Podman socket provides support for the Docker v1.40 API, which meets
s2i's needs. The logic looks for the podman socket in the following
order:

1. Rootless podman socket
2. Root podman socket
3. Docker socket

If none are found, the default falls back to the docker socket. As
before, the `--url` and `DOCKER_HOST` environment variables can
override the container engine socket location.

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>